+++ src/mockfn/mock.clj
@@ -2,6 +2,8 @@
   (:require [mockfn.matchers :as matchers]
             [mockfn.parser]))
 
+(defrecord Calling [func])
+
 (defn- matches-arg?
   [[expected arg]]
   (if (satisfies? matchers/Matcher expected)
@@ -24,16 +26,22 @@
 (defn- unexpected-call [func args]
   (format "Unexpected call to %s with args %s." func args))
 
-(defn- return-value-for
+(defn- get-value-for
   [func spec args]
   (when (-> spec :return-values (for-args args) #{::unexpected-call})
     (throw (ex-info (unexpected-call func args) {})))
   (-> spec :times-called (for-args args) (swap! inc))
   (-> spec :return-values (for-args args)))
 
+(defn- base-value-or-invoke [func spec args]
+  (let [mocked-value (get-value-for func spec args)]
+    (if (instance? Calling mocked-value)
+      (apply (:func mocked-value) args)
+      mocked-value)))
+
 (defn mock [func spec]
   (with-meta
-    (fn [& args] (return-value-for func spec (into [] args)))
+    (fn [& args] (base-value-or-invoke func spec (into [] args)))
     spec))
 
 (defn- doesnt-match [function args matcher times-called]