+++ test/mockfn/examples/basic_usage.clj
@@ -18,6 +18,12 @@
       (is (= :result-1 (one-fn :argument-1)))
       (is (= :result-2 (one-fn :argument-2)))))
 
+  (testing "providing - cause function to throw exception"
+    (mfn/providing [(one-fn :argument-1) (mfn/calling #(throw (ex-info "one-fn mocked to fail" {:arg %})))
+                    (one-fn :argument-2) :result-2]
+      (is (thrown? ExceptionInfo (one-fn :argument-1)))
+      (is (= :result-2 (one-fn :argument-2)))))
+
   (testing "providing with more than one function"
     (mfn/providing [(one-fn :argument) :result-1
                     (other-fn :argument) :result-2]